pyproject.toml
[project]
name = "client-project"
version = "0.1.0"
description = "ML Platform Client Project Example"
requires-python = ">=3.11"
dependencies = [
    "mlflow>=2.10.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "pytest>=7.4.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
